Glaze support and Default skills (formatting, gtest and json) (#48)
### Why 
At twig we are beginning to support glaze. Having a skill for it will be
greatly appreciated.
Furthermore the choice to go for default implementation for formatting,
json etc is that it greatly simplifies the usage of stronk. At the same
time automatically created stronk types (like derived units) will not be
able to print and not require users to either specialize or write custom
functions. It is still possible to opt out of the behavior, either by
not including the headers that implement the behavior, or by custom
implementation of the functions

### What was changed
- Add a `can_glaze_de_and_serialize` skill
- Glaze requires atleast clang-15 so we need to remove clang-13,
clang-14 support
- Added support for default formatting with std::format
- Default support for fmt::format (can still be specialized)
- Default support for gtest (breaking change, skill has been removed)
- Default support for nlohmann json (breaking change, skill has been
removed)